refactor: 패키지 이동 및 config 분리(JPAAuditing 분리하지 않으면 테스트 시 어노테이션 추가 필요)
