+++ src/main/java/com/catcher/datasource/config/JpaAuditingConfig.java
@@ -1,0 +1,10 @@
+package com.catcher.datasource.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+
+}
